// src/store/user/user-types.ts
export enum USER_ACTION_TYPES {
  SET_CURRENT_USER = "SET_CURRENT_USER",
  CHECK_USER = "CHECK_USER",
  GOOGLE_SIGNIN = "GOOGLE_SIGNIN",
  EMAIL_SIGNIN = "EMAIL_SIGNIN",
  SIGNIN_SUCCESS = "SIGNIN_SUCCESS",
  SIGNIN_FAILED = "SIGNIN_FAILED",
  SIGNUP = "SIGNUP",
  SIGNUP_SUCCESS = "SIGNUP",
  SIGNUP_FAILED = "SIGNUP",
  SIGNOUT = "SIGNOUT",
  SIGNOUT_SUCCESS = "SIGNOUT_SUCCESS",
  SIGNOUT_FAILED = "SIGNOUT_FAILED",
}